Cleanup file structure and prepare for tests
