+++ lib/src/services/globalConfig.ts
@@ -3,24 +3,7 @@ import appContext from "../utils/appContext";
 import fs from "fs";
 import yaml from "js-yaml";
 import { z } from "zod";
-
-/**
- * Creates a file with t
- * @param filename
- * @param defaultContents
- */
-function createFileIfMissing(filename: string, defaultContents?: any) {
-  const dir = path.dirname(filename);
-
-  // create the directory if it doesn't already exist
-  if (!fs.existsSync(dir)) fs.mkdirSync(dir);
-
-  // create the file if it doesn't already exist
-  if (!fs.existsSync(filename)) {
-    // create the file, writing the `defaultContents` if provided
-    fs.writeFileSync(filename, defaultContents || "", "utf-8");
-  }
-}
+import { createFileIfMissing } from "../utils/fileSystem";
 
 const ZGlobalConfigYAML = z.record(
   z.string(),
@@ -33,45 +16,11 @@ const ZGlobalConfigYAML = z.record(
 
 type GlobalConfigYAML = z.infer<typeof ZGlobalConfigYAML>;
 
-const ZProjectConfigYAML = z.object({
-  projects: z
-    .array(
-      z.object({
-        id: z.string(),
-      })
-    )
-    .optional(),
-  variants: z
-    .array(
-      z.object({
-        id: z.string(),
-      })
-    )
-    .optional(),
-  outputs: z.array(
-    z.object({
-      format: z.enum(["i18next"]),
-    })
-  ),
-});
-
-type ProjectConfigYAML = z.infer<typeof ZProjectConfigYAML>;
-
-export const DEFAULT_PROJECT_CONFIG_JSON: ProjectConfigYAML = {
-  projects: [],
-  variants: [],
-  outputs: [
-    {
-      format: "i18next",
-    },
-  ],
-};
-
 /**
  * Read data from a global config file
- * @param {string} file defaults to `CONFIG_FILE` defined in `constants.js`
- * @param {*} defaultData defaults to `{}`
- * @returns { GlobalConfigYAML }
+ * @param file defaults to `CONFIG_FILE` defined in `constants.js`
+ * @param defaultData defaults to `{}`
+ * @returns
  */
 export function readGlobalConfigData(
   file = appContext.configFile,
@@ -94,6 +43,12 @@ function writeGlobalConfigData(file: string, data: object) {
   fs.writeFileSync(file, yamlStr, "utf8");
 }
 
+/**
+ * Save a token to the global config file
+ * @param file The path to the global config file
+ * @param hostname The hostname to save the token for
+ * @param token The token to save
+ */
 export function saveToken(file: string, hostname: string, token: string) {
   const data = readGlobalConfigData(file);
   data[hostname] = []; // only allow one token per host
